+++ tests/test_SOAPifySOAP.py
@@ -1,58 +1,65 @@
+import pytest
 import SOAPify
 import numpy
 from numpy.testing import assert_array_equal
 import h5py
 import MDAnalysis as mda
 import HDF5er
-from testSupport import getUniverseWithWaterMolecules
+from .testSupport import getUniverseWithWaterMolecules
+import pytest
 
 
-def test_MultiAtomicSoapify():
+@pytest.fixture(
+    scope="module",
+    params=[
+        None,
+        ["O"],
+    ],
+)
+def fixture_AtomMask(request):
+    return request.param
+
+
+def test_MultiAtomicSoapify(fixture_AtomMask, engineKind_fixture):
     nMol = 27
     u = getUniverseWithWaterMolecules(nMol)
-    HDF5er.MDA2HDF5(u, "testH2O.hdf5", "testH2O", override=True)
+    fname = f"testH2O{''.join([i for i in fixture_AtomMask]) if fixture_AtomMask else ''}.hdf5"
+    HDF5er.MDA2HDF5(u, fname, "testH2O", override=True)
     n_max = 4
     l_max = 4
     rcut = 10.0
-    with h5py.File("testH2O.hdf5", "a") as f:
+    with h5py.File(fname, "a") as f:
         soapGroup = f.require_group("SOAP")
         trajGroup = f["Trajectories/testH2O"]
         SOAPify.saponify(
-            trajGroup, soapGroup, rcut, n_max, l_max, useSoapFrom="dscribe"
+            trajGroup,
+            soapGroup,
+            rcut,
+            n_max,
+            l_max,
+            useSoapFrom=engineKind_fixture,
+            SOAPatomMask=fixture_AtomMask,
         )
         assert soapGroup["testH2O"].attrs["n_max"] == n_max
         assert soapGroup["testH2O"].attrs["l_max"] == l_max
         assert "O" in soapGroup["testH2O"].attrs["species"]
         assert "H" in soapGroup["testH2O"].attrs["species"]
         assert numpy.abs(soapGroup["testH2O"].attrs["r_cut"] - rcut) < 1e-8
-        assert "centersIndexes" not in soapGroup["testH2O"].attrs
+        if fixture_AtomMask == None:
+            assert "centersIndexes" not in soapGroup["testH2O"].attrs
+        else:
+            assert_array_equal(
+                soapGroup["testH2O"].attrs["centersIndexes"],
+                [i * 3 for i in range(nMol)],
+            )
         assert (
             soapGroup[f"testH2O"].shape[-1]
             == (1 + l_max) * n_max * n_max
             + 2 * (1 + l_max) * ((n_max + 1) * n_max) // 2
         )
 
-        soapGroup = f.require_group("SOAPOxygen")
-        SOAPify.saponifyGroup(
-            f["Trajectories"],
-            soapGroup,
-            10.0,
-            n_max,
-            l_max,
-            SOAPatomMask=["O"],
-            useSoapFrom="dscribe",
-        )
-        assert soapGroup["testH2O"].attrs["n_max"] == n_max
-        assert soapGroup["testH2O"].attrs["l_max"] == l_max
-        assert "O" in soapGroup["testH2O"].attrs["species"]
-        assert "H" in soapGroup["testH2O"].attrs["species"]
-        assert numpy.abs(soapGroup["testH2O"].attrs["r_cut"] - rcut) < 1e-8
-        assert_array_equal(
-            soapGroup["testH2O"].attrs["centersIndexes"], [i * 3 for i in range(nMol)]
-        )
 
-
-def test_MultiAtomicSoap():
+def test_MultiAtomicSoapifyGroup(engineKind_fixture):
     nMol = 27
     u = getUniverseWithWaterMolecules(nMol)
     HDF5er.MDA2HDF5(u, "testH2O.hdf5", "testH2O", override=True)
@@ -62,7 +69,12 @@ def test_MultiAtomicSoap():
     with h5py.File("testH2O.hdf5", "a") as f:
         soapGroup = f.require_group("SOAP")
         SOAPify.saponifyGroup(
-            f["Trajectories"], soapGroup, rcut, n_max, l_max, useSoapFrom="dscribe"
+            f["Trajectories"],
+            soapGroup,
+            rcut,
+            n_max,
+            l_max,
+            useSoapFrom=engineKind_fixture,
         )
         assert soapGroup["testH2O"].attrs["n_max"] == n_max
         assert soapGroup["testH2O"].attrs["l_max"] == l_max
@@ -84,7 +96,7 @@ def test_MultiAtomicSoap():
             n_max,
             l_max,
             SOAPatomMask=["O"],
-            useSoapFrom="dscribe",
+            useSoapFrom=engineKind_fixture,
         )
         assert soapGroup["testH2O"].attrs["n_max"] == n_max
         assert soapGroup["testH2O"].attrs["l_max"] == l_max